# Expression: faster equals and hash