@Bouncner 2020-03-31
Do you have an idea what the impact is on SF 1.0?