@mrks 2020-03-31
Way more than expected, but then again my trust in these numbers has lately been challenged.

```diff
 +----------------+--------------------+-------+--------------------+-------+------------+---------+
 | Benchmark      | prev. iter/s       | runs  | new iter/s         | runs  | change [%] | p-value |
 +----------------+--------------------+-------+--------------------+-------+------------+---------+
 | TPC-H 01       | 1.2217390537261963 | 74    | 1.2723581790924072 | 77    | +4%        |  0.0000 |
+| TPC-H 02       | 74.69950866699219  | 4482  | 78.99553680419922  | 4741  | +6%        |  0.0000 |
 | TPC-H 03       | 9.226232528686523  | 554   | 9.318384170532227  | 560   | +1%        |  0.0000 |
 | TPC-H 04       | 8.136442184448242  | 489   | 8.461881637573242  | 508   | +4%        |  0.0000 |
 | TPC-H 05       | 4.224071979522705  | 254   | 4.271946430206299  | 257   | +1%        |  0.0001 |
+| TPC-H 06       | 256.9529724121094  | 15418 | 272.342041015625   | 16341 | +6%        |  0.0000 |
 | TPC-H 07       | 9.498988151550293  | 570   | 9.73809814453125   | 585   | +3%        |  0.0006 |
 | TPC-H 08       | 7.387296676635742  | 444   | 7.587373733520508  | 456   | +3%        |  0.0000 |
 | TPC-H 09       | 1.7512286901474    | 106   | 1.757331371307373  | 106   | +0%        |  0.3192 |
 | TPC-H 10       | 3.355991840362549  | 202   | 3.416388750076294  | 205   | +2%        |  0.0782 |
 | TPC-H 11       | 38.4580078125      | 2308  | 40.37691116333008  | 2423  | +5%        |  0.0000 |
 | TPC-H 12       | 16.7611026763916   | 1006  | 17.100431442260742 | 1027  | +2%        |  0.0000 |
 | TPC-H 13       | 2.651763677597046  | 160   | 2.632814407348633  | 158   | -1%        |  0.0000 |
 | TPC-H 14       | 44.38142013549805  | 2663  | 43.78355026245117  | 2628  | -1%        |  0.0000 |
 | TPC-H 15       | 117.6553955078125  | 7060  | 120.279052734375   | 7217  | +2%        |  0.0000 |
 | TPC-H 16       | 9.640525817871094  | 579   | 9.763609886169434  | 586   | +1%        |  0.0000 |
 | TPC-H 17       | 5.335076808929443  | 321   | 5.551656723022461  | 334   | +4%        |  0.0000 |
+| TPC-H 18       | 1.2211766242980957 | 74    | 1.2950519323349    | 78    | +6%        |  0.0000 |
 | TPC-H 19       | 9.815033912658691  | 589   | 10.097067832946777 | 606   | +3%        |  0.0000 |
 | TPC-H 20       | 31.11349105834961  | 1867  | 32.30149841308594  | 1939  | +4%        |  0.0000 |
 | TPC-H 21       | 1.3779687881469727 | 83    | 1.4036909341812134 | 85    | +2%        |  0.0001 |
 | TPC-H 22       | 18.688142776489258 | 1122  | 18.300582885742188 | 1099  | -2%        |  0.0000 |
 | geometric mean |                    |       |                    |       | +2%        |         |
 +----------------+--------------------+-------+--------------------+-------+------------+---------+
```