@mrks 2020-03-31
For Q18, this is questionable as there are only four potential variations of that query, so most of the optimizations should be cached.