@Bouncner 2020-04-02
> For Q18, this is questionable as there are only four potential variations of that query, so most of the optimizations should be cached.

So this expression change does not occur during execution?