@mrks 2020-04-02
Mmmh, some operators use expressions. To be honest, I don't feel like diving too deep into this. The benchmarks were meant as a sanity check to make sure we don't break anything.